Added books to dashboard overview -> can now also be deleted by admin. Updated home page with a header for styling purposes. Completely changed, but also completed search bar on home page -> can now look up book titles or authors. Started on filtering on genre; will work further on tomorrow.

+++ app/Http/Controllers/DashboardController.php
@@ -2,19 +2,25 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Book;
 use App\Models\Review;
 use App\Models\User;
+use Gate;
 use Illuminate\Http\Request;
 
 class DashboardController extends Controller
 {
     public function show(User $user) {
         $users = User::all();
-        return view('dashboard', compact('users'));
+        $books = Book::all();
+
+        return view('dashboard', compact('users', 'books'));
     }
 
     public function destroy(User $user)
     {
+        Gate::authorize('admin-access');
+
         $user->delete();
         return redirect()->route('dashboard');
     }
